Update patch set 1
Patch Set 1:

The context of this is back in June, I submitted a patch to LKML to add support for handing REBOOT2 arguments via the Bootloader Control Block so that things like "adb reboot recovery" work well on generic Intel Architecture. There's no special registers to flip, I just update the BCB and SYSLINUX has been modified to know where to look:

https://patchwork.kernel.org/patch/1133431/

This ended up in a discussion between Colin and Greg KH where Greg claimed that use of the REBOOT2 API in this case is inappropriate and this should be moved to userspace. The thread more or less ended without any resolution. So I am trying to do it in userspace instead by making a small modification to android_reboot().

In order for me to move this to userspace, I need for processes to be able to modify the BCB in the 'misc' partition directly; 0600 perms won't work. I had been trying to change the perms in init, but then ran into the problem where the chmod/chown commands didn't follow the symlinks for the block device nodes. I wanted to change the perms for 'misc' partition node to 0660.

As it turns out, the situation is even more complex since  both Settings and adbd need to have the ability to reboot into an alternate target, so I may have to go back to doing this in the kernel anyway and just maintain the patch locally. (there's currently no overlap in the groups that adbd and Settings belong to that I can find.) But this change seems like it could be generally useful so I sent it here.

Patch-set: 1