Update patch set 1
Patch Set 1:

All of it can be hard-coded at build time. But we are trying to see about not hard-coding stuff in the build when we can.

Our group is trying to move device configuration parameters from being build-time decisions to install- or run-time decisions wherever possible. In this way a single binary image can run on multiple devices, with some baseline level of quality, without having to modify it.

As an example, the device's physical camera configuration is specified through ro.camera.* variables. For getting camera(s) to work on development boards, the installation is interactive, the user would specify which camera is front, back, etc. The config gets set in the property file created in the installation partition. (I'm aware there's other camera related stuff set in XML configs, this is just for example).

Other work in this area (which we have not upstreamed yet) includes modifying ueventd so that it can automatically load kernel modules at runtime based on uevents, like udevd.

It's unlikely that we'll achieve product-quality experience for all our dev boards with a single image, but these efforts have already made our lives a lot easier from a distro perspective, as with these changes we can build an image which runs "pretty good" on any particular dev board we run it on.

Patch-set: 1