Update patch set 6
Patch Set 6: Code-Review-1

Properties are a limited resource, global to the system, and serve a specific purpose. Piggy backing on that for device uevents feels awkward to me and makes me uneasy about the bloat. This to me just looks like an abuse of properties. Keep in mind that properties cannot be deleted once created, so your event trigger property will live on forever until reboot. For example, we previously had a problem where eventhub was abusing properties for a similar purpose and we eventually ran out of space for property storage and the device became wonky.

As asked before, could you provide a full example of how it was used? Like for the squashfs mount, what was it doing? Why wasn't vold used for doing the mount?

Patch-set: 6
Reviewer: Gerrit User 1000413 <1000413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1