Update patch set 6
Patch Set 6:

On the installation side, sorry to not be authoritative but I understand that the reason vold was not used was that the installation device was not known apriori so several matching rules/mounts were being used with the device-trigger/property-trigger approach. I suppose vold would ignore any errors also so perhaps this wouldn't be an issue. I'm assuming squashfs is or could be made to be supported in vold.

Anyway the motivation for all this definitely was the PnP scenarios that Srinivas refers to. We just took advantage of it for installation.

As far as property exhaustion, I haven't envisioned a scenario where many are used. I thought you were going to rewrite the property system anyway. :-)

Patch-set: 6