allow adb host to ignore specific USB devices
The way adb scans for USB devices can interfere with USB power
management commands run on that device - specifically powering on/off
the device. This change allows a user to pass a list of devices
to "blacklist" via an environment variable to adb.

An example script that was affected by adb is:

  http://people.linaro.org/~doanac/sdmux/sdmux.sh

If adb was run in between the power off and power on commands, then
the device could not be powered on and mounted properly.

Change-Id: I97b24b46c3f65f56eead5ce59a49d4ab3f624657
Signed-off-by: Andy Doan <andy.doan@linaro.org>