Update patch set 2
Patch Set 2:

shit. this is another change that's gone missing. the reason you see the conflict is because internal master has the right fix for this, which is to remove the files in question:

commit 663005248dd5d62198feb2188bc9e841739b673e
Author: Elliott Hughes <enh@google.com>
Date:   Tue Sep 11 11:18:38 2012 -0700

    The localtime_tz and mktime_tz extensions are now in bionic.
    
    Bug: 7012465
    
    (cherry-pick of d03072ffde3279fbb60591a0d07c820d8e792762.)
    
    Conflicts:
    
        libcutils/Android.mk
    
    Change-Id: I5e34edd4c8f59cc0a4d5dcc4740707ae063608a0

this is particularly crazy because the change in question was originally committed to AOSP: https://android-review.googlesource.com/#/c/42743/

Patch-set: 2