Update patch set 6
Patch Set 6:

> we eventually ran out of space for property storage and the device became wonky.

I wasn't aware of that. Before my time. Had I known I would have pushed in a different direction. The property was convenient as it allowed grouping multiple devices to a single trigger.

> ... and suggested use properties instead.

See my Aug 28 21:03 comment in which I don't recommend starting/stoping a service on device insertion/removal. Because it is assync and by the time the service starts the device might already be gone or going away, which implies that the service already needs to handle device removal. At which point it might was well handle device re-insertion.

Patch-set: 6