Update patch set 10
Patch Set 10:

(20 comments)

This is definitely a huge clash in style. Commented on some things inline, the patch will follow in a day or two. However,  let me explain my error handling pattern that you find "ew" (backwards go to)
While I agree that the backwards goto is a gross idiom, when used to replace loop constructs. When used in this case, it is really nice, notice my functions have a single exit point.

You can think of my pattern as a crappy version of try,catch,and finally (although not as good as this: https://bitbucket.org/billcroberts/cexception), let me explain:
1. You encounter some error, you throw it by setting rc and goto err
2. the err label is the catch
3. the finally is the out

That's my justification for my approach.

Like I said, Ill go your way on all of this (sans using int8_t for write as that's a correctness issue), as its style and doesn't hurt correctness; but I think I died a little bit on the inside :-P

Patch-set: 10