Update patch set 10
Patch Set 10:

The two exit points doesn't address the case when you always want some code to be run on exit of the function. You either need to duplicate the code, or make a function, which in all essence takes up as much space as a jmp.

Either way, your project not mine :-P I will even go to spaces . Give me a few days to clean it up, and thoroughly test it.

Patch-set: 10