Update patch set 10
Patch Set 10:

I don't care at all about a few instructions, I care about readability and maintainability.  The pattern you describe is fine when that is what the code needs and the extra complexity makes sense, but everywhere I've seen it so far is when it would be trivial to use a simpler pattern.  Use a single "goto out" when you want the error handling to use the same exit path as the return (for example when holding a lock that should be released), or a "goto err" that jumps to after the normal exit path if the error path should undo something (for example freeing allocated memory before returning NULL).

Patch-set: 10
Label: Code-Review=0