Update patch set 13
Patch Set 13: Code-Review-1

(26 comments)

with regards to the libaudit stuff, is there an upstream for this?  Otherwise libaudit.h <-> netlink.c still doesn't make sense to me from a file organization perspective.

Also, are you trying to keep this in line with any other upstream auditd's, or is this just going to be on it's own?

style is still all over the place.

Patch-set: 13
Reviewer: Gerrit User 1012391 <1012391@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1