Update patch set 13
Patch Set 13:

The layout was originally kept the same to the GPL'd interface. But we broke that with the return -errno change, so I agree and think that netlink.c can go to libaudit.c

This auditd is not trying to stay in line with anything upstream.

Patch-set: 13
Label: Verified=0