Signing ability in mkbootimg
Author: Andrew Boie <andrew.p.boie@intel.com>
Author: Jianxun Zhang <jianxun.zhang@intel.com>

Add two new arguments to sign boot image
--signsize: size of signature
--signexec: command to sign boot image. It is expected to take
unsigned payload from STDIN when mkbootimage invokes it and output
signature ONLY to STDOUT.

The signature block is appended page-aligned to the boot image
right after the 2nd-stage loader. The size of the signature block
is stored in one of the previously unused fields and is 0 if
the boot image is not signed, maintaining backward compatibility.

As an example, one can use openssl for this purpose:

BOARD_MKBOOTIMG_ARGS := --signsize 256 \
        --signexec "openssl dgst -sha256 -sign $(TARGET_BOOT_IMAGE_KEY)"

Change-Id: I50c42f5cd44b1bf92575d14d865c59ec6a3af512
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>