Update patch set 1
Patch Set 1:

That is what security_compute_create() is doing - it is computing the security context for the new process based on the security context of the executable.  But the problem is which argument represents the executable; in the non-logwrapper case, it is the arg0 while in the logwrapper case it is the first non-option argument to logwrapper.
I don't think we want to call security_compute_create on every argument and assume that the first one that yields a different context than the caller is the right one, as it is specific to logwrapper that the argument is an executable.
If this approach won't fly, then I guess we go back to the one suggested by Bill, i.e. specify seclabel in the init.rc and allow entrypoint for the logwrapper executable (but only for the daemon domains,not for all domains).

Patch-set: 1
Label: Verified=0