Update patch set 1
Patch Set 1:

There is an alternative approach to implementing this functionality that just uses the NETLINK_SELINUX socket as another socket to poll in the main event loop rather than checking the status page on each received uevent.  I can upload that approach as well as a separate change if desired, but am unclear as to whether it would be preferred.

Patch-set: 1