Added net.dns1 prop for emulator, as Android emulator ignores it by default.
This commit sets up the system property which is actually used by the Connectivity Service (frameworks/base/services/java/com/android/server/ConnectivityService.java). It fixes an (obsolete?) convention where the dns was affected directly by the interface (i.e. "net.eth0.dns1=10.0.2.3"), which causes the Android Emulator (goldfish) to ignore this value, and effectively have no DNS resolving at all.
An immediate fix can be either add reference to net.eth%s.dns%s in the ConnectivityService and possibly on the dhcp code as well which would be bloated, or just stick to the apparant new convention.
I chose the latter as a one line fix which gets the job done.

Change-Id: Id4364129e9a82c1f48403068a837aca54de07944