Update patch set 1
Patch Set 1:

What's the warning?  A comma at the end of the enum list is nice, it means you don't have to edit 2 lines to add a new enum to the end of the list, and makes the diffs cleaner.  In this case, I doubt anything will ever get added to the end of the list, so I'd be fine with this patch if its necessary, but I'd like to investigate disabling the warning first.

Patch-set: 1