Fix utf16_to_utf8_length to detect invalid strings
utf16_to_utf8_length is used to measure the number
of bytes needed for utf8 conversion, typically to
get the necessary size of the buffer to send to utf16_to_utf8.

However, when an invalid surrogate pair is encountered
the size is underestimated compared to what utf16_to_utf8
will actually write, which may lead to buffer overflows.

The measurement function has been changed to return -1 on
invalid pairs (just like utf32_to_utf8_length) and unit tests
are added.

Change-Id: I568bbdb065d7181d18d5661be2e62aace50e607f