Update patch set 3
Patch Set 3:

We've never really had a program where we had a single bootloader support multiple boards. Everyone always keeps a per-device bootloader branch because the bootloader almost always has to do something different for every board. But looking at some of the options you want to pass, it doesn't seem like it's the right thing for the bootloader to do.

Have you considered using device tree? The device tree for the specific device would then have all the configuration necessary, and you would just have to flash the right device tree onto the device or have the bootloader choose the right device tree that's appended to your kernel, which is what we do here for newer devices. Is there x86 support for device tree?

I suppose there's not really a big issue in merging this patch since it doesn't interfere with anything, but I think forcing everything on the command line through this image doesn't sound an optimum way to do it and puts the kernel options specifics far away from the kernel and into the Android build where it doesn't belong (my main concern).

Patch-set: 3