Allow emulator to set ro.boot.selinux property on kernel cmd line
This fix will allow the emulator to set selinux to enforcing, permissive
or disabled via the ro.boot.selinux property.

An example emulator command line would be:
emulator -qemu -append selinux=permissive -show-kernel \
-kernel $ANDROID_BUILD_TOP/kernel/goldfish/arch/arm/boot/zImage

It allows the SELinux mode to be set early in the init process as using
"-prop ro.boot.selinux=permissive" on the emulator command line does not
work, neither does setting "setprop ro.boot.selinux permissive" in
"on early-init" of init.rc as the property is initialised too late in
the init process to be actioned.

Change-Id: I4c17e39ae81cc2a6a3bb3203a622d224d9ea6727
Signed-off-by: Richard Haines <richard_c_haines@btinternet.com>