Update patch set 1
Patch Set 1:

As Elliott mentioned, I'm in the midst of switching the stack unwinding code to use libunwind. I have not added full x86-64 support, but that should only be a matter of modifying the makefile in external/libunwind.

The current sticking point is that the unwind code did not properly detect when the stack was done unwinding. If you would like to test the libunwind version, modify the file:

system/core/libbacktrace/Android.mk

And add the target x86 or x86_64 to the list right below the line that begins:

# To enable using libunwind on each arch, add it to the list below.

If you could help to fix the libunwind code, that would be great.

Patch-set: 1
Label: Code-Review=0