Update patch set 4
Patch Set 4:

I'm ok with stripping it from user builds, just like su is. I just didn't know how invasive of a change would be tolerable.

But then when you build eng vs user, this could be something that provides inconsistencies between the builds, and their are already a lot of those.

Patch-set: 4