AArch64: Use negative values in error enumeration
With AArch64 the enumerations with values in the range 0x80000000 to
0x80000007 are being assigned to ssize_t typed variables which are
64-bit rather than 32-bit, and are being used in conditions where they
are being checked for being negative. This is not the case when
ssize_t is  64-bit, so for __LP64__ redefine these as being actual
negative values. Starting with an arbitrary value that must be larger
than the largest errno value.

Change-Id: I7a031a940a28658b3bf34bebac93dfb3ba397b05
Signed-off-by: Marcus Oakland <marcus.oakland@arm.com>