Update patch set 1
Patch Set 1:

guangzhu: do we _know_ that libunwind causes the slowdown? if so, we can commit this until cferris has chance to look into the libunwind problem. if we still don't know whether it's libunwind's fault, can you cherry-pick this and see if you still have trouble with your test?

Patch-set: 1