Pass -a command-line option to ADB server.
ADB is used to provide a bridge between (Gfx Debug) debugger
and a GLink server running on an Android target.

For this to work, ADB must be able to accept TCP/IP connections from
any machine, not just the local host.  In JellyBean, it was necessary
to patch the ADB code to get this to work, but in KitKat there is a
new "-a" command-line option to ADB which allows this.  However, when
ADB spawns a new instance of itself as a server process, it does not
pass on the "-a" option, so the server will not accept connections
from other machines.

This change fixes this problem by passing the "-a" option to the
server process on its command line.

Change-Id: If6742d313fe29792487a1d82ad6ce2a39321ce9f
Author: rstuckey <richard.stuckey@intel.com>
Signed-off-by: Bo Huang <bo.b.huang@intel.com>
Signed-off-by: Zhuang,Jin Can <jin.can.zhuang@intel.com>