Update patch set 1
Patch Set 1:

The BatteryManager status is only for charging purposes, and "USB" specifies a maximum current limited by SDP (500mA for USB 2.x), which I'd rather not indicate in this case.  UsbManager is normally used for detecting enumerated devices for host communication, does this not work for the cases you're interested in?

Patch-set: 1