Update patch set 1
Patch Set 1:

There's a risk of a slowdown, and we've tried hard to avoid forcing a relabeling of all of /data, which in the past has been expensive. So far, none of the relabeling we've done has had a noticeable impact on boot time.

Instead, the focus has been on identifying small subsets of /data which could benefit from relabeling, and adding the restorecon_recursive call.

I've been meaning to write a small init service which would do targeted relabeling on the first boot after a system update, rather than on every boot. That way, we would only take the hit one time, and only for directories where the SELinux label changed. Haven't had time to write it yet...  :-(

Patch-set: 1