Increase try_get_word robustness
is_readable_map() does not garantee that any read access to this area
will succeed: File map or device map could be only partially associated
with real memory. This will lead to SIGSEGV or SIGBUS errors.
The purpose of this patch is to limit register retrieval to only stack
regions.

Change-Id: Ice49d6c9fc959ee003991ffcfab6e4e76d0af8ea
Signed-off-by: Emmanuel Berthier <emmanuel.berthier@intel.com>
Signed-off-by: hanpengx <pengfeix.han@intel.com>