restorecon:  Only call setfilecon if the context is not already correct.
We already have this optimization in the libselinux and toolbox
restorecon code.  I think it was omitted from the init built-in
because originally we only applied it to individual files (not
recursively) and only when we knew the context would not be correct
in the first place.  Regardless, getfilecon() should be less
expensive than setfilecon() even when setting the same value,
and the common case should be that the file has the right context
already.

Change-Id: Icc1dcf6968429b7519aafea4972ed422eb8f413d
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>