WIP Move debuggerd to C++.
Making all of debuggerd compile in C++ and removing the calls to the
C libbacktrace interface. This is the first step to completely remove the
C libbacktrace interface, which is kind of hacky.

The next step is to remove the maps info structure and create a class to
represent the data. This is to allow certain optimizations that would be
extremely difficult to implement using the C interface.

Change-Id: I026f3f823f591dc838263470b24373e96eb4a505