Update patch set 3
Patch Set 3:

the existing APPLE cruft was because of art too. i added that back when we were being told we had to support Mac OS. so that shouldn't be taken as precedent. it should be removed, not extended.

in this particular restricted case, if the runtime is crashing, the best thing to do on Mac OS is to just call abort(3). the Mac has a built-in debuggerd equivalent that generates its own dumps that are much better than ours. you can view them in Console.app.

libbacktrace's layer of indirection is between multiple _working_ backends. this change adds a non-functional one, which isn't useful.

even if someone were to add real general Mac OS support, it wouldn't look much like this anyway.

Patch-set: 3