Update patch set 1
Patch Set 1:

Sorry it took so long to respond. I haven't had a chance to look at this closely. Both the x86 and mips code use the same dwarf code, which should be finding the eh_frame.

Specifically, the code in external/libunwind/src/dwarf/Gfind_unwind_table.c in the dwarf_find_unwind_table function is where the eh_frame debug information is supposed to be found. The PT_GNU_EH_FRAME case should be hit for mips, if not then that's the problem.

Patch-set: 1