Update patch set 1
Patch Set 1:

We just use USB vs. AC charger as a hint as to whether it's likely charging is capped at 500mA max or may be charging at a faster rate.  I'd prefer to keep the AC designation for faster rates.  It's true the terminology is inaccurate and the UI can be misleading in a number of ways, hopefully to be improved some day.

Patch-set: 1