Only apply restorecon_recursive once per OTA/build.
For any persistent directory (e.g. /data, /persist), we only want
to apply restorecon_recursive once per OTA/build to fix the security contexts.
Avoid repeatedly walking the directory tree on each boot by setting
a security.restorecon_last xattr on each directory during a
restorecon_recursive tree walk to the value of ro.build.date.utc
and skipping the traversal if the xattr is already set and matches
the current value of ro.build.date.utc.

For /sys, the attempt to set the xattr will fail but this
is harmless.

This required loading the build properties earlier during boot,
prior to post-fs-data where the restorecon_recursive calls are performed.
There is still an issue with restorecon_recursive /persist calls
which are in the fs section of the init.*.rc files.
We cannot move the loading of build props prior to that
since that is where /system is mounted IIUC.  We could move
the restorecon_recursive call to post-fs-data if nothing
accesses /persist prior to that time.

I am unclear on the charger case.  I loaded build props in that
case to maintain consistency, as they would have been loaded by
the subsequent property_service_init.  But I am not sure that
this ever works since system is not mounted until the charger
actions, at least on hammerhead.

A getrestoreconlast toolbox command is included to help with seeing
the effects of the change, but this can be dropped from the final version.

Change-Id: If6215753dbd520031d052918eb3c1a0685196fb7
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>