Only apply restorecon_recursive once to any persistent directory.
For any persistent directory (e.g. /data, /persist), we only want
to apply restorecon_recursive once to fix the security contexts.
Avoid repeatedly walking the directory tree again by checking for
and creating a .restorecon_done flag file at the top of the
directory tree.

For /sys, the attempt to create the flag file will fail but this
is harmless.

Change-Id: If6215753dbd520031d052918eb3c1a0685196fb7
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>