Update patch set 2
Patch Set 2:

You bring up a good point about the service based approach. If this is a problem, shouldn't it already be a problem for ueventd, which is started in init.rc's early-init section? Couldn't we start a fixup service in the post-fs section?

Colin: Is there anyway to make a service start call block until it's complete?

Patch-set: 2
Label: Verified=0