Update patch set 2
Patch Set 2:

Why wouldn't just checking the top level directory against the expected value work all the time? Then there wouldn't be a need for a .restorecon_done file at all.

If we were to do that, then we couldn't rely on a restorecon of a parent directory to always affect the child. We'd have to start doing something like:

restorecon_recursive /data/misc
restorecon_recursive /data/misc/tzinfo
...

In my mind, it's still cleaner to go with a /sefixups / /data/sefixups style approach and integrate it into fs_mgr like Colin suggested. But I don't want to over engineer an unnecessary solution if this approach has the potential to work...

Patch-set: 2