Update patch set 2
Patch Set 2:

That would mean that we would re-run all of the restorecon_recursive commands on every change to ro.build.fingerprint, i.e. on every build?  Whereas we typically only need to do it once per directory tree.  For example, /data/mediadrm only needs to be restorecon'd once on an upgrade from any existing release of Android to master or the next version of Android.
The getxattr/setxattr calls would always fail on sysfs and we always want to fully restorecon sysfs anyway, so they are just adding overhead for that one.

Patch-set: 2