Update patch set 2
Patch Set 2:

Long term, I'd like to do a relabeling of /data on every OTA. 

init.rc: restorecon_recursive /data

If we change an SELinux label on a subdirectory of /data, merely checking the label on /data itself wouldn't be sufficient. However, checking the fingerprint allows us to realize that *something* might have changed, forcing the relabel.

Re: extra blocks. At least from my experience modifying ext4_utils, the space in the inode is REALLY small, and right now it's mostly filled with the SELinux label. This will almost certainly require an extra block on the disk. I *thought* the kernel did caching and reference counting of xattrs, so it may be possible for the kernel to reuse the xattr block for multiple directories. Worst case is that we'd have 1 extra block used per directory in /data (if we decide to relabel all of /data, that is).

Patch-set: 2