log: add formats for printing zero-padded pointers in hex
Many places in Android's code make use of "08x" or "08X" to print
pointers with zero-padding. This is fine on 32bit systems but leads
to a truncated display on 64bit systems if building at all.

This commit introduces formats that will allow to fix these issues
in an elegant way. As they aren't defined by the C standard they
didn't belong to inttypes.h.

Change-Id: Ifc1e6675afbb36737412a7736f4f104ca6fad6b3
Signed-off-by: Kévin PETIT <kevin.petit@arm.com>