Update patch set 3
Patch Set 3:

Is there a compile flag you can use that only does this checking on eng and userdebug builds? This would not resolve the side-effects issue (which should be nil, really ... the developers must make sure of this). This would resolve the additional runtime processing issue for releases?

Patch-set: 3
Label: Verified=0