HACK to get gralloc.omap3.so working on newer Android
Use a hack from the XDA forum to get this old gralloc.omap3.so
running on newer versions of Android:

http://forum.xda-developers.com/showpost.php?p=44326678&postcount=264

What happened is that the android::CallStack API changed, with
the result that gralloc.omap3.so no longer links against libutils.
The poster worked around this problem by including the old version
of android::CallStack but renaming it to android::CallStac2, and
then using a binary editor to s/CallStack/CallStac2/g in
gralloc.omap3.so.

Change-Id: I65c836461445ca7be51dc53f093691c8ae8063b6