Update patch set 4
Patch Set 4:

Alright, that rules out any databases. The only viable candidate was the SettingsProvider one but that is writable by any app with a sig|system perm (and it's a separate app). I'll go ahead and dump it to a file. IIRC, didn't Bill at one point have some file that contained the hashes of these files. Is any of that relevant again or is there some code to steal.

Patch-set: 4