Update patch set 4
Patch Set 4: -Code-Review

I don't think you need anything fancy, just read the file if it exists and compare with the current hash of seapp_contexts; if not equal, then call restorecondatadata and when it finishes, write out the new hash to the file.  The file doesn't need to contain anything other than the hash value itself.

Patch-set: 4