[WIP] Add a config for dual mode zygote.
Renamed from init.zygote64.rc, because we will not
need a 64 only configuration (except for testing, but
even there it creates more issues than it solves).

The life cycles of both zygotes are controlled by init,
and the assumption here is that they will be available
always.

If we move to a model where one zygote execs the other,
then we can revert this change.

Change-Id: Id0be001ea6f934c3c2022d89a63aae9fae66cc38