Update patch set 2
Patch Set 2:

@Stephen, regarding stripping the logcat prefix:

# adb logcat -v raw -b main > logcat.txt

Glad to hear that msg= was not a requirement, want to preserve in-memory space. I also wanted (but left as-is) to strip the timestamp out (leaving just the sequence number) as the time is also recorded on the entries (currently I use time of reception rather than time-of-record). This would trim the in-memory size of the entries. If we could strip the timestamp out of the message string and ausearch still worked, that would be a bonus.

Ideas?

Patch-set: 2