Update patch set 2
Patch Set 2:

Don't know why your ausearch isn't correctly handling comm or path; on Fedora it seems to work correctly.
Anyway, the only concern with omitting the redundant fields is that if we have to generate raw records to use audit2allow and then go back and try to debug some denial, we'll have to keep both raw and full logcat output and correlate them, right?  audit2allow itself shouldn't care about any of the fields except for the ones needed to generate the rule, but it does have a -v option or similar that extracts the other info and embeds it as a comment line.

Patch-set: 2
Reviewer: Gerrit User 1010099 <1010099@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0