Update patch set 4
Patch Set 4:

Thanks Steve for bringing up proctitle kernel patches.

Excuse me for my ignorance on what's really going on here. I saw some comments of stripping the timestamp. Isn't it used as the fingerprint of audit records? So in the future if Google or OEMs enable syscall auditing to retrieve proctitle they won't be able to corelate the records together as 1 event?

Patch-set: 4
Label: Verified=0