Update patch set 4
Patch Set 4:

I mostly just wanted ausearch -if to work sufficiently to permit trivial decoding of hex-encoded values and the like, although one could certainly do that with a script instead.
ausearch -if has other issues, e.g. the host version on Fedora doesn't know how to decode the syscall number for ARM.
For correlation, the serial number is generally enough for a set of audit records in "close proximity"; the serial number should be unique until rollover, so augmenting it with the timestamp is only necessary for looking at a large set of audit events over time.
The -v time info from logd will be more meaningful than the kernel's notion of time reported in the audit message, particularly for Android.

Patch-set: 4