Update patch set 1
Patch Set 1:

If your only problem is setting permissions on sysfs files, that is already supported in ueventd.  It's a little painful, because you have to find the canonical node for the inserted device (the path without any symlinks in it), but it might work for you.  See https://android.googlesource.com/device/samsung/manta/+/master/ueventd.manta.rc for an example.

Patch-set: 1